delete unnecessary usings, create MassTransit extension methods
