+++ Shared/EventBus.Messages/Constants/EventBusConstant.cs
@@ -1,14 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Shared.EventBus.Messages.Constants
+﻿namespace Shared.EventBus.Messages.Constants
 {
-	public static class EventBusConstants
-	{
-		public const string BasketCheckoutQueue = "ticketstatusqueue-queue";
-		public const string EmailWithChangesQueue = "emailwithchanges-queue";
-	}
+    public static class EventBusConstants
+    {
+        public const string BasketCheckoutQueue = "ticketstatusqueue-queue";
+        public const string EmailWithChangesQueue = "emailwithchanges-queue";
+    }
 }
