+++ Shared/EventBus.Messages/Enums/Status.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Shared.EventBus.Messages.Enums
+﻿namespace Shared.EventBus.Messages.Enums
 {
     public enum Status
     {
