+++ Shared/EventBus.Messages/Events/GetTicketStatusEvent.cs
@@ -1,9 +1,4 @@
 ﻿using Shared.EventBus.Messages.Enums;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Shared.EventBus.Messages.Events
 {
