+++ Shared/EventBus.Messages/Events/UpdatedInfoEvent.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Shared.EventBus.Messages.Events
+﻿namespace Shared.EventBus.Messages.Events
 {
     public class UpdatedInfoEvent : IntegrationBaseEvent
     {
