+++ Shared/MassTransit/Filters/PublishLoggingFilter.cs
@@ -1,10 +1,5 @@
 ﻿using MassTransit;
 using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Shared.MassTransit.Filters
 {
