+++ src/Services/Catalog/Catalog.Domain/Specification/ConcertSpecifications/ConcertsWithDate.cs
@@ -1,9 +1,4 @@
 ﻿using Catalog.Domain.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Catalog.Domain.Specification.ConcertSpecifications
 {
