+++ src/Services/Catalog/Catalog.Domain/Specification/TicketsSpecifications/TicketsWithStatus.cs
@@ -1,10 +1,5 @@
 ﻿using Catalog.Domain.Entities;
 using Catalog.Domain.Enums;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Catalog.Domain.Specification.TicketsSpecifications
 {
