+++ src/Services/Catalog/Catalog.WebApi/Extensions/ServiceExtensions.cs
@@ -13,11 +13,8 @@ using Catalog.Infrastructure.Services;
 using FluentValidation;
 using FluentValidation.AspNetCore;
 using Microsoft.EntityFrameworkCore;
-using Shared.EventBus.Messages.Constants;
-using Shared.MassTransit.Filters;
 using StackExchange.Redis;
 using System.Reflection;
-using MassTransit;
 
 namespace Catalog.WebApi.Extensions
 {
@@ -57,35 +54,6 @@ namespace Catalog.WebApi.Extensions
             services.AddGrpc(opt =>
             {
                 opt.EnableDetailedErrors = true;
-            });
-
-            services.AddMassTransit(x => 
-            {
-                var assembly = Assembly.GetAssembly(typeof(TicketStatusConsumer));
-                x.AddConsumers(assembly);
-
-                string host = config["RabbitMQ:Host"];
-                var virtualHost = config["RabbitMQ:VirtualHost"];
-                var username = config["RabbitMQ:Username"];
-                var password = config["RabbitMQ:Password"];
-
-                x.UsingRabbitMq((context, cfg) =>
-                {
-                    cfg.Host(host, virtualHost, h =>
-                    {
-                        h.Username(username);
-                        h.Password(password);
-                    });
-
-                    cfg.UseConsumeFilter(typeof(ConsumeLoggingFilter<>), context);
-                    cfg.UsePublishFilter(typeof(PublishLoggingFilter<>), context);
-
-                    cfg.ReceiveEndpoint(EventBusConstants.BasketCheckoutQueue, x =>
-                    {
-                        //x.Bind<GetTicketStatusEvent>();
-                        x.ConfigureConsumer<TicketStatusConsumer>(context);
-                    });
-                });
             });
 
             return services;
