+++ src/Services/Catalog/Catalog.WebApi/appsettings.json
@@ -16,8 +16,5 @@
     "VirtualHost": "/",
     "Username": "guest",
     "Password": "guest"
-    //"ReceiveEndpoints": {
-    //  "UserCreated": "words_user-created_queue"
-    //}
   }
 }
