+++ src/Services/Identity/Identity.Application/Dtos/MailDto/MailSettingsDto.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Identity.Application.Dtos.MailDto
+﻿namespace Identity.Application.Dtos.MailDto
 {
     public class MailSettingsDto
     {
