+++ src/Services/Identity/Identity.Application/Interfaces/IMailService.cs
@@ -1,11 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using MimeKit;
-using MailKit;
-using Identity.Application.Dtos.MailDto;
+﻿using Identity.Application.Dtos.MailDto;
 
 namespace Identity.Application.Interfaces
 {
