+++ src/Services/Identity/Identity.Application/Services/MailService.cs
@@ -1,16 +1,16 @@
 ﻿using Identity.Application.Dtos.MailDto;
-using Identity.Application.Interfaces;
 using MailKit.Security;
 using Microsoft.Extensions.Options;
 using MimeKit;
-using MailKit;
 using MailKit.Net.Smtp;
+using Identity.Application.Interfaces;
 
 namespace Identity.Application.Services
 {
-    public class MailService : Interfaces.IMailService
+    public class MailService : IMailService
     {
         private readonly MailSettingsDto _mailSettings;
+
         public MailService(IOptions<MailSettingsDto> mailSettings)
         {
             _mailSettings = mailSettings.Value;
@@ -18,23 +18,19 @@ namespace Identity.Application.Services
 
         public async Task SendMailAsync(MailDataDto mailData, CancellationToken ct = default)
         {
-            // Initialize a new instance of the MimeKit.MimeMessage class
             var mail = new MimeMessage();
-
-            // Sender
             mail.Sender = MailboxAddress.Parse(_mailSettings.Mail);
+
             foreach (string mailAddress in mailData.ToMail)
             {
                 mail.To.Add(MailboxAddress.Parse(mailAddress));
             }
 
-            // Receiver
             foreach (string mailAddress in mailData.ToMail)
             {
                 mail.To.Add(MailboxAddress.Parse(mailAddress));
             }
 
-            // Add Content to Mime Message
             var body = new BodyBuilder();
             mail.Subject = mailData.Subject;
             body.HtmlBody = mailData.Body;
