+++ src/Services/Identity/Identity.WebApi/appsettings.json
@@ -25,8 +25,5 @@
     "VirtualHost": "/",
     "Username": "guest",
     "Password": "guest"
-    //"ReceiveEndpoints": {
-    //  "UserCreated": "words_user-created_queue"
-    //}
   }
 }
