+++ src/Services/Order/Order.WebApi/appsettings.json
@@ -12,16 +12,10 @@
   "GrpcServer": {
     "Address": "https://localhost:5046"
   },
-  //"EventBusSettings": {
-  //  "HostAddress": "http://localhost:15672"
-  //},
   "RabbitMQ": {
     "Host": "localhost",
     "VirtualHost": "/",
     "Username": "guest",
     "Password": "guest"
-    //"ReceiveEndpoints": {
-    //  "UserCreated": "words_user-created_queue"
-    //}
   }
 }
